reinstated cache

+++ prereq/install_nest.sh
@@ -6,27 +6,27 @@ set -e  # stop execution in case of errors
 export NEST_VERSION="2.10.0"
 export NEST="nest-$NEST_VERSION"
 pip install cython
-#if [ ! -f "$HOME/$NEST_VERSION/configure" ]; then
+if [ ! -f "$HOME/$NEST_VERSION/configure" ]; then
     wget https://github.com/nest/nest-simulator/releases/download/v$NEST_VERSION/$NEST.tar.gz -O $HOME/$NEST.tar.gz;
     pushd $HOME;
     tar xzf $NEST.tar.gz;
     popd;
-#else
-#    echo 'Using cached version of NEST sources.';
-#fi
+else
+    echo 'Using cached version of NEST sources.';
+fi
 mkdir -p $HOME/build/$NEST
 pushd $HOME/build/$NEST
-#if [ ! -f "$HOME/build/$NEST/config.log" ]; then
+if [ ! -f "$HOME/build/$NEST/config.log" ]; then
     export VENV=`python -c "import sys; print sys.prefix"`;
     echo "VENV: $VENV"
     $HOME/$NEST/configure --with-mpi --prefix=$VENV;
     make;
-#else
-#    echo 'Using cached NEST build directory.';
-#    echo "$HOME/$NEST";
-#    ls $HOME/$NEST;
-#    echo "$HOME/build/$NEST";
-#    ls $HOME/build/$NEST;
-#fi
+else
+    echo 'Using cached NEST build directory.';
+    echo "$HOME/$NEST";
+    ls $HOME/$NEST;
+    echo "$HOME/build/$NEST";
+    ls $HOME/build/$NEST;
+fi
 make install
 popd
